Reduce RuboCop offences

+++ controllers/users_controller.rb
@@ -26,26 +26,35 @@ module ShowTracker
       redirect '/', error: t('errors.no_such_show') if show.nil?
 
       usershow = Usershow.for_user_and_show current_user.id, show.id
-      redirect '/users/my-shows', error: t('errors.not_in_my_shows') if usershow.nil?
+
+      if usershow.nil?
+        flash[:error] = t('errors.not_in_my_shows')
+        redirect '/users/my-shows'
+      end
 
       usershow.destroy
 
-      redirect '/users/my-shows', success: t('successes.removed_show', name: show.name)
+      flash[:success] = t('successes.removed_show', name: show.name)
+      redirect '/users/my-shows'
     end
 
     get '/decrement-episode/:usershow_id', auth: :logged do
       usershow = Usershow.with_id params[:usershow_id]
 
       if usershow.episode.zero? || usershow.episode.nil?
-        redirect '/users/my-shows', error: 'You haven\'t watched any episodes yet!'
+        flash[:error] = 'You haven\'t watched any episodes yet!'
+        redirect '/users/my-shows'
       end
 
       usershow.decrement_episode
       usershow.save
 
-      flash[:success] = t('successes.episode_unwatched', season: usershow.season,
-                                                        episode: usershow.episode + 1,
-                                                           show: usershow.show.name)
+      flash[:success] = t(
+        'successes.episode_unwatched',
+        season: usershow.season,
+        episode: usershow.episode + 1,
+        show: usershow.show.name
+      )
 
       redirect '/users/my-shows'
     end
@@ -55,9 +64,12 @@ module ShowTracker
       usershow.increment_episode
       usershow.save
 
-      flash[:success] = t('successes.episode_watched', season: usershow.season,
-                                                      episode: usershow.episode,
-                                                         show: usershow.show.name)
+      flash[:success] = t(
+        'successes.episode_watched',
+        season: usershow.season,
+        episode: usershow.episode,
+        show: usershow.show.name
+      )
       redirect '/users/my-shows'
     end
 
@@ -71,8 +83,11 @@ module ShowTracker
       usershow.decrement_season
       usershow.save
 
-      flash[:success] = t('successes.season_unwatched', season: usershow.season + 1,
-                                                          show: usershow.show.name)
+      flash[:success] = t(
+        'successes.season_unwatched',
+        season: usershow.season + 1,
+        show: usershow.show.name
+      )
       redirect '/users/my-shows'
     end
 
@@ -81,8 +96,11 @@ module ShowTracker
       usershow.increment_season
       usershow.save
 
-      flash[:success] = t('successes.season_watched', season: usershow.season,
-                                                        show: usershow.show.name)
+      flash[:success] = t(
+        'successes.season_watched',
+        season: usershow.season,
+        show: usershow.show.name
+      )
       redirect '/users/my-shows'
     end
 
@@ -118,20 +136,22 @@ module ShowTracker
     post '/signup' do
       username         = params[:username]
       password         = params[:password]
-      confirm_password = params[:confirm_password]
+      # confirm_password = params[:confirm_password]
       email            = params[:email]
       first_name       = params[:first_name]
       last_name        = params[:last_name]
 
       password_salt = hash_salt
-      password_hash = hash_pasword(params[:password], password_salt)
+      password_hash = hash_pasword(password, password_salt)
 
-      user = User.create(username: username,
-                         password: password_hash,
-                         salt: password_salt,
-                         email: email,
-                         first_name: first_name,
-                         last_name: last_name)
+      User.create(
+        username: username,
+        password: password_hash,
+        salt: password_salt,
+        email: email,
+        first_name: first_name,
+        last_name: last_name
+      )
 
       flash[:success] = t('successes.register')
       redirect NAMESPACE + '/login'
